implemented getUserBalances
